Updates detection of already-running docker compose instances
Current parsing of textural docker compose ps output assumes `"--------"` will delimit output. This is no longer the case. This PR amends us to use the JSON output formatter.

* Adds PsWithJson method to DockerCompose.cs to use within DockerCompose.cs `CheckIfRunning` method. This protects us somewhat from changes to the text output of docker compose ps.

* Updates tests.

I did consider adding a JSON parser library to parse the output (and would be happy to switch to that), but I did not want to bloat any consuming code bases.